ft-#66: Rename files and identifiers
