--- src/pages/Auth/ForgetPassword/type.ts
@@ -1,6 +1,0 @@
-type TForgetLoginPasscode = {
-  phone: string;
-  email: string;
-};
-
-export type { TForgetLoginPasscode };
