+++ src/pages/Auth/ForgotLoginPasscode/type.ts
@@ -1,0 +1,6 @@
+type TForgetLoginPasscodeFormValues = {
+  phone: string;
+  email: string;
+};
+
+export type { TForgetLoginPasscodeFormValues };
